Extract PresentationToolbar and add button to show recent files
